% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ALF_cat}
\alias{ALF_cat}
\title{Function to concatenate all gene sequences into one sequence for each species}
\usage{
ALF_cat(ALF_sim)
}
\arguments{
\item{ALF_sim}{An "ALFsim" object containing the sequence to be concatenated}
}
\value{
An "ALFsim" object with concatenated sequences
}
\description{
Function to concatenate all gene sequences into one sequence for each species
}

