% Generated by roxygen2 (4.0.0): do not edit by hand
\name{load_ALF}
\alias{load_ALF}
\title{Load sequences and trees from an ALF simulation into R}
\usage{
load_ALF(simdir, which_specs = NULL, return_tree = TRUE,
  true_align = TRUE)
}
\arguments{
\item{simdir}{The directory where the simulation output you wish to load is located. This is returned by
\code{\link{run_ALF}}.}

\item{which_specs}{A vector of integers expressing which species sequences you wish to load}

\item{return_tree}{Boolean. Returns the underlying tree if TRUE, only the sequences if FALSE}

\item{true_align}{Boolean. Should the function load the true alignement from the simulation. Currently not used.}
}
\value{
An "ALFsim" object containing the requested simulated sequences and the simulated tree
(if return_tree = TRUE).
}
\description{
Load sequences and trees from an ALF simulation into R
}

