% Generated by roxygen2 (4.0.0): do not edit by hand
\name{run_ALF}
\alias{run_ALF}
\title{a function to create an ALF configuration file (*.drw)}
\usage{
run_ALF(simname, nspec, ngenes, mingenelen, PAMunits, brate, drate, indelrate,
  dir, ALF_dir, capture_prog = FALSE)
}
\arguments{
\item{simname}{Name for the simulation to be simulated}

\item{nspec}{Number of species to simulate}

\item{ngenes}{Number of genes to simulate in the genome}

\item{mingenelen}{Minimum length of a gene under the simulation}

\item{PAMunits}{Distance from root to tips on simulated phylogeny, in PAM units: time it takes
to accumulate 1 changed amino acid out of every 100 amino acids (codon based evolutionary unit)}

\item{brate}{Birth rate for Birth-Death phylogeny simulation}

\item{drate}{Death rate for Birth-Death phylogeny simulation}

\item{indelrate}{Rate at which insertions and deletion are incorporated into genomes}

\item{dir}{Directory where the simulated genome should be stored}

\item{ALF_dir}{Directory containing the \code{alfsim} executable file}

\item{capture_prog}{Should the progress of ALF be sent to stdout (so it can be capture by \code{knitr} for example)?}
}
\value{
A character vector containing the path to the directory in which the ALF simulation saves its output.
}
\description{
a function to create an ALF configuration file (*.drw)
}

